requirements.txt
fastapi
motor
uvicorn[standard]
pymongo[srv]
openai
python-dotenv